// src/contracts/types/exercise.ts
export interface components {
    schemas: {
// ...
        ExerciseCreation: {
            code: string;
            title: string;
            description?: string;
        };
// ...
        Exercise: {
            /** Format: uuid */
            exerciseId: string;
            code: string;
            title: string;
            description?: string;
            status: components["schemas"]["ExerciseStatus"];
            currentVersion?: {
                /** Format: uuid */
                versionId: string;
                versionNumber: number;
                /** Format: date-time */
                publishedAt?: string;
            };
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt?: string;
        };
// ...
    };
// ...
}